bot_backy/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
openai>=1.14.3

flask>=3.0.2
flask-cors>=4.0.0

requests>=2.31.0

# for IMDb API access (if needed)
IMDbPY

# for fuzzy search
fuzzywuzzy[speedup]

# for command-line parsing (some utils use it)
typer

# for backups with progress bars or logs
tqdm